+++ trustfulIP/auth/src/main/java/com/lee/security/JwtAuthenticationFilter.java
@@ -42,7 +42,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
             String token = header.substring(7);
 
             try {
-                Integer userId = JwtUtil.getUserIdInt(token);
+                Long userId = JwtUtil.getUserIdInt(token);
 
                 User user = userMapper.selectOne(
                         new QueryWrapper<User>().eq("userId", userId)
